Interface Input management
1) Context:
I want to decouple input and display from the current hat I have to build the key on top.

To achieve this I will split the input and display in 2 different abstract models.

2) Changeset:
Add:
- interface.input: definitions
- interface.input.pico_lcd_1_3: implementation

3) Tests:
None - Trivial

4) Related Documentation: (_)No/(Y)es/(P)ending review
(N) Require a documentation change to README
(Y) Merge to mainline (if not why?)